+++ apps/backyard/src/app.routes.tsx
@@ -1,5 +1,5 @@
-import { LoaderFunction, redirect, RouteObject } from 'react-router-dom';
-import { auth, collection } from 'lib';
+import { auth } from 'lib';
+import { redirect, RouteObject } from 'react-router-dom';
 
 import AppShell from './app/app-shell';
 
@@ -18,20 +18,6 @@ import MatchesView from './app/matches';
 import ProfileView from './app/profile';
 import TipsView from './app/tips';
 
-const rootLoader: LoaderFunction = async () => {
-  // Load small users collection early and eager
-  const users = await collection('users').get();
-
-  // Auth state now sure, so check
-  if (auth.currentUser === null) {
-    return redirect('/login');
-  }
-
-  return {
-    users,
-  };
-};
-
 const appRoutes: RouteObject[] = [
   {
     path: '/login',
@@ -40,8 +26,11 @@ const appRoutes: RouteObject[] = [
   {
     path: '/',
     element: <AppShell />,
-    id: 'root',
-    loader: rootLoader,
+    loader: async () => {
+      if (!auth.currentUser) {
+        return redirect('/login?from=' + window.location.pathname);
+      }
+    },
     children: [
       { index: true, element: <Dashboard /> },
       { path: 'turnier', element: <ChampionshipView /> },