+++ apps/backyard/src/model/user.ts
@@ -1,5 +1,0 @@
-import { User as FirebaseUser } from 'firebase/auth';
-export type User = Pick<
-  FirebaseUser,
-  'uid' | 'email' | 'displayName' | 'photoURL'
->;