+++ apps/backyard/src/states/auth-state.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+import { Profile } from '@/model/profile';
+
+export const authState = atom<Profile | null>({
+  key: 'authState',
+});