+++ apps/backyard/src/states/profile-state.ts
@@ -1,4 +1,0 @@
-import { Profile } from '@/model/profile';
-import { atom } from 'jotai';
-
-export const profileState = atom<Profile | null>(null);