+++ apps/foh/CHANGELOG.md
@@ -1,5 +1,12 @@
 # foh
 
+## 0.0.25
+
+### Patch Changes
+
+- Updated dependencies
+  - lib@0.0.11
+
 ## 0.0.24
 
 ### Patch Changes